feat(login): improve kakao login callback handling and add login page
